+++ python/src/agent_api/types/__init__.py
@@ -2,6 +2,7 @@ from agent_api.types.common import *
 from agent_api.types.input import *
 from agent_api.types.responses import *
 from agent_api.types.streaming import *
+from agent_api.types.memories import *
 from agent_api.types.skills import *
 from agent_api.types.tools import *
 from agent_api.types.volumes import *